Invoice system and many more
This commit has a whole new invoice system as well as many bugfixes and a prototype .html file that will be used to add items to the beverages even outside the app. (Mono is a but buggy with json handling dlls, so if you can't compile, just use the .exe on windows or navigate to the bin folder on linux and type 'mono *.exe' into the terminal. Going to fix this in the next commit.)